Handle user-profile when running as admin.
In an installation for a non-admin user, an R installation by default
will be installed into the local users' home directory. However, we
still need administrative privileges to add the binding link for ArcGIS
10.3.1. So, when the process is 'Run as Administrator', the registry
HKCU reflects the admin, not the user who has R installed.

This change explicitly finds the user profile, even in an escalated
installation case, and makes sure we can identify the R install in
either case. This fix is unnecessary for 10.4 and later releases.