Hotfix for 3.2.4revised release
Currently the hook on the ArcGIS side doesn't work with a release that
contains more than one character, so anyone upgrading to the latest
version who uses a global package directory won't be able to run
scripts. To work around this for affected versions, use the registry
hookup that we have exposed to ensure users' can still access R, albiet
in a way that requires the bridge to update.