expose r_set_install, writing R registry values
Allow users to update the default R to use "Current Version" via the
registry. This setting is an optional setting on the installation of R,
but in its absense, we have to guess which version of R is the desired
one to use for the binding package. Instead, provide a mechanism to
update these settings if missing.

Also allows version information to be obtained either through the
registry, or any way possible, in order to detect the current registry
settings.